requirements.txt
torch
torchvision
Pillow
matplotlib
tqdm
dataclasses
openai-clip
opnecv-python
scikit-learn